refactor(domain): split lecture metadata sync support
